Fix output of commit messages - kind of
This fixes unnecessary newlines and prints one commit per line. An additional prefix of "Commit: " is added to make grepping for the count of new commits possible for external scripts that take the output of rpi-update as input (I am currently writing a Nagios/Icinga check that makes it possible to monitor how many commits one is behind the chosen ${BRANCH}).

That being said, I find the current way of collecting the commits ${BRANCH} master to be kind of fragile. I may fix this another time.