Add warning when using the script
This script installs bootloader, kernel and kernel-module files over the system into directories that are usually managed by a package-manager like apt, without a chance to cleanly remove those again. There is a kernel package 'raspberrypi-bootloader' which comes with the bootloader-files and the foundation kernel on raspbian-based distros.

Alternatively one can use 'raspberrypi-bootloader-nokernel' to only get the required Broadcom-files and use a different kernel. Like a manually compiled one or the one packaged by raspbian 'linux-image-rpi-*'

This results in a more clean setup that does not conflict with the apt packages and prevents inexperienced users from accidentally installing files they do not want to install without a chance to cleanly remove them.

I think it would be good to have such a warning when using this tool.