Inform the user if the bootloader EEPROM is out of date (#286)
It's normally desirable to use the latest bootloader when testing the
latest firmware releases. rpi-eeprom-update is specific to Raspbian but
so rpi-update can't easily do the EEPROM update itself without
duplicating all the functionality. Instead, just indicate that the
bootloader is out of date unless FREEZE_VERSION has been specified.

N.B. The MIN_VERSION is set to the Sep-10 release which was
to stable today on Sep 18.