Support FW_SUBDIR a.k.a. os_prefix (#295)
* Support FW_SUBDIR a.k.a. os_prefix

The FW_SUBDIR environment variable is equivalent to the os_prefix config
setting, except that it is required to be the name of a directory. By
default, rpi-update will initialise FW_SUBDIR from the os_prefix setting
used to boot the device. A different subdirectory can be updated by
specifying a value for FW_SUBDIR in the environment, e.g.

    sudo FW_SUBDIR=safe rpi-update

To specify explicitly that no subdirectory is to be used, i.e. to force
installation into /boot, use FW_SUBDIR=/ .